+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 Kemono uses [semantic versioning](https://semver.org/).
 
 ### v1.4
-- New "Semibooru" UI
+- New UI
+- Search posts
+- Pagination
+- Implemented thumbnail generation
 - General code rewrites and cleanup
+- Fixed issue causing error when ID is not in the lookup database
+- Slightly longer API cache times
 
 ### v1.3.1
 - Replaced Cloudscraper module